3_agentic_runtime_level2/requirements.txt
python-dotenv>=1.0.1
pydantic>=2.8.2
openai>=1.40.0
